+++ segno/utils.py
@@ -118,22 +118,6 @@ def matrix_to_lines(matrix, x, y, incby=1):
             last_bit = 0x0
 
 
-def matrix_with_border_iter(matrix, version, border):  # pragma: no cover
-    """\
-    Returns an interator / generator over the provided matrix which includes
-    the border.
-
-    :param matrix: An iterable of bytearrays.
-    :param int version: A version constant.
-    :param int border: The border size or ``None`` to specify the
-            default quiet zone (4 for QR Codes, 2 for Micro QR Codes).
-    :raises: py:exc:`ValueError` if an illegal border value is provided
-    """
-    import warnings
-    warnings.warn('Use utils.matrix_iter', DeprecationWarning)
-    return matrix_iter(matrix, version, border=border)
-
-
 def matrix_iter(matrix, version, scale=1, border=None):
     """\
     Returns an interator / generator over the provided matrix which includes