+++ tests/test_png.py
@@ -214,5 +214,5 @@ def _get_png_info(**kw):
 
 
 if __name__ == '__main__':
-    pytest.main(['-x', __file__])
+    pytest.main([__file__])