+++ tests/test_svg.py
@@ -641,5 +641,5 @@ def svg_as_matrix(buff, border):
 
 
 if __name__ == '__main__':
-    pytest.main(['-x', __file__])
+    pytest.main([__file__])