+++ tests/test_writers.py
@@ -69,4 +69,4 @@ def test_get_writable_not_stream2():
 
 
 if __name__ == '__main__':
-    pytest.main(['-x', __file__])
+    pytest.main([__file__])